+++ src/components/board/task/task.js
@@ -53,7 +53,7 @@ export default class Task {
   }
 
   getElement() {
-    if(!this._element) {
+    if (!this._element) {
       this._element = createElement(this.getTemplate());
     }
     return this._element;
