--- src/components/board/taskslist.js
@@ -1,8 +1,0 @@
-import createTask from "./task/task.js";
-
-const createTasksList = (tasks) => tasks.reduce((tasksList, task) => {
-   tasksList += createTask(task);
-   return tasksList;
-}, ``);
-
-export default createTasksList;
